requirements.txt
streamlit
sqlite3
pandas
plotly
bcrypt
openpyxl
Pillow